+++ src/mariner/mixins.py
@@ -43,9 +43,9 @@ class ComparableMixin:  # pylint: disable=too-few-public-methods
 
 
 class GetPageMixin:  # pylint: disable=too-few-public-methods
-    """Represt a web scraper."""
+    """Represent a web scraper."""
 
-    async def get(self, url: str) -> str:
+    async def get(self, url: str, headers: str=None, cookies: str=None) -> str:
         """Get the requested page.
 
         Args:
@@ -55,6 +55,6 @@ class GetPageMixin:  # pylint: disable=too-few-public-methods
             Raw HTML page.
         """
         with async_timeout.timeout(10):
-            async with aiohttp.ClientSession() as session:
+            async with aiohttp.ClientSession(headers=headers, cookies=cookies) as session:
                 async with session.get(url) as response:
                     return await response.text()