+++ CHANGELOG.md
@@ -10,7 +10,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Support for Nyaa.si tracker.
 - Support for Limetorrents tracker.
 - Set download path for downloading torrent files.
-- Automatically get working proxy ThePirateBay, KickAssTorrents, LimeTorrents.
+- Automatically get working proxy for ThePirateBay, LimeTorrents.
+
+### Fixed
+- Fix support for KickAssTorrents.
 
 ## [1.0.3] - 2018-03-04
 ### Fixed