+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [1.1.0] - 2018-05-16
 ### Added
 - Support for Nyaa.si tracker.
 - Support for Limetorrents tracker.
@@ -82,3 +82,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [1.0.1]: https://gitlab.com/radek-sprta/mariner/compare/v1.0.0...v1.0.1
 [1.0.2]: https://gitlab.com/radek-sprta/mariner/compare/v1.0.1...v1.0.2
 [1.0.3]: https://gitlab.com/radek-sprta/mariner/compare/v1.0.2...v1.0.3
+[1.1.0]: https://gitlab.com/radek-sprta/mariner/compare/v1.0.3...v1.1.0